// src/config/enum/config-key.enum.ts
export enum ConfigKey {
  // Twitch
  TWITCH_USERNAME = 'TWITCH_USERNAME',
  TWITCH_TOKEN = 'TWITCH_TOKEN',
  TWITCH__OAUTH_CLIENT_ID = 'TWITCH__OAUTH_CLIENT_ID',
  TWITCH__OAUTH_CLIENT_SECRET = 'TWITCH__OAUTH_CLIENT_SECRET',
  TWITCH__OAUTH_REDIRECT_URI = 'TWITCH__OAUTH_REDIRECT_URI',

  // Database
  DATABASE_USER = 'DATABASE_USER',
  DATABASE_NAME = 'DATABASE_NAME',
  DATABASE_PASSWORD = 'DATABASE_PASSWORD',
  DATABASE_HOST = 'DATABASE_HOST',
  DATABASE_PORT = 'DATABASE_PORT',
}